+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Enh #39: Adapt to `yiisoft/yii-runner` of version `^2.0` (@vjik)
 - New #40: In the `HttpApplicationRunner` constructor make parameter "environment" optional (`null` by default) and add
-- new parameters that set config group names for container, bootstrap and events (@vjik)
+  new parameters that set config group names for container, bootstrap and events (@vjik)
 
 ## 1.1.2 November 10, 2022