+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Yii Runner HTTP Change Log
 
-## 1.2.0 under development
+## 2.0.0 under development
 
 - Enh #39: Adapt to `yiisoft/yii-runner` of version `^2.0` (@vjik)
 - New #40: In the `HttpApplicationRunner` constructor make parameter "environment" optional, default `null` (@vjik)